Refactored and did some DRYing
