--- lib/stradivari/dislocated.rb
@@ -1,13 +1,0 @@
-module Stradivari::Dislocated
-  module Tabs
-
-    autoload :Generator, 'stradivari/dislocated/tabs/generator'
-
-  end
-
-  module Content
-
-    autoload :Generator, 'stradivari/dislocated/content/generator'
-
-  end
-end
